root/opt/leds/requirements.txt
openrgb-python==0.3.1
